+++ amdsmi_cli/amdsmi_parser.py
@@ -1437,6 +1437,7 @@ class AMDSMIParser(argparse.ArgumentParser):
 
         # Help text for Arguments only on Guest and BM platforms
         metrics_help = "Metric XGMI information"
+        xgmi_source_status_help = "Source GPU XGMI Link information"
         xgmi_link_status_help = "XGMI Link Status information"
 
         # Create xgmi subparser
@@ -1447,6 +1448,7 @@ class AMDSMIParser(argparse.ArgumentParser):
 
         # Optional Args
         xgmi_parser.add_argument('-m', '--metric', action='store_true', required=False, help=metrics_help)
+        xgmi_parser.add_argument('-s', '--source-status', action='store_true', required=False, help=xgmi_source_status_help)
         xgmi_parser.add_argument('-l', '--link-status', action='store_true', required=False, help=xgmi_link_status_help)
 
         # Add Universal Arguments